requirements.txt
alembic==1.13.2
annotated-types==0.7.0
certifi==2024.8.30
cffi==1.17.1
curl_cffi==0.7.1
fake-useragent==1.5.1
greenlet==3.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
rich==13.8.1
selectolax==0.3.21
SQLAlchemy==2.0.35
typing_extensions==4.12.2
